Refactor system information class
