+++ graphenex/core/web/providers.py
@@ -1,10 +1,10 @@
 from graphenex.core.web import auth_socketio, socketio, emit
-from graphenex.core.utils.helpers import Information
+from graphenex.core.utils.sysinfo import SysInformation
 
 
 @socketio.on('get_network_info')
 @auth_socketio
 def send_network_info():
-    masks = Information.get_network()
+    masks = SysInformation.get_network_info()
     emit('get_network_info', masks)
