+++ graphenex/core/web/views.py
@@ -2,7 +2,8 @@
 # -*- coding: utf-8 -*-
 
 from graphenex.core.web import app, logger, socketio
-from graphenex.core.utils.helpers import check_os, get_modules, mod_json_file, Information
+from graphenex.core.utils.helpers import check_os, get_modules, mod_json_file
+from graphenex.core.utils.sysinfo import SysInformation
 
 from flask import render_template, session, request, redirect, flash
 from flask_socketio import emit, disconnect
@@ -60,7 +61,7 @@ def main():
         'index.html',
         title="grapheneX [Web]",
         mod_count=get_mod_count(module_dict),
-        sys_info_full=Information.get_all())
+        sys_all_info=SysInformation.get_all_info())
 
 
 @app.errorhandler(404)
